refactor: split gateway feedback surfaces
