+++ crates/dcc-mcp-cli/src/presentation/mod.rs
@@ -1,5 +1,6 @@
 pub mod cli;
 pub(crate) mod components_cmd;
+pub(crate) mod feedback_cmd;
 pub(crate) mod marketplace_cmd;
 pub mod output;
 pub(crate) mod update_cmd;
