+++ crates/dcc-mcp-gateway/src/gateway/handlers/feedback.rs
@@ -65,6 +65,92 @@ pub async fn handle_v1_feedback(
         .into_response()
 }
 
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use axum::body::Body;
+    use axum::http::Request;
+    use tower::ServiceExt;
+
+    use crate::gateway::handlers::rest_impl::rest_impl_tests::{response_json, test_gateway_state};
+
+    #[tokio::test]
+    async fn gateway_feedback_records_with_zero_live_dcc_instances() {
+        let gateway = test_gateway_state("1.2.3");
+        assert!(gateway.live_instances_async().await.is_empty());
+        let mut updates = gateway.events_tx.subscribe();
+        let app = crate::gateway::router::build_gateway_router(gateway.clone());
+        let request_body = json!({
+            "tool_name": "houdini.ui_control__act",
+            "intent": "Open the render menu",
+            "attempt": "Invoked the semantic menu action",
+            "blocker": "The owning DCC process exited",
+            "severity": "blocked",
+            "dcc_type": "houdini",
+            "instance_id": "aaaaaaaa-0000-0000-0000-000000000000",
+            "request_id": "request-42",
+            "job_id": "job-42"
+        });
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .method("POST")
+                    .uri("/v1/feedback")
+                    .header(axum::http::header::CONTENT_TYPE, "application/json")
+                    .body(Body::from(serde_json::to_vec(&request_body).unwrap()))
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        let (status, body) = response_json(response).await;
+
+        assert_eq!(status, StatusCode::CREATED);
+        assert_eq!(body["ok"], true);
+        assert_eq!(body["event_resource_uri"], GATEWAY_EVENTS_URI);
+        assert!(uuid::Uuid::parse_str(body["feedback_id"].as_str().unwrap()).is_ok());
+
+        let events = gateway.event_log.recent_events(1);
+        assert_eq!(events.len(), 1);
+        assert_eq!(events[0].event, EventKind::FeedbackReported);
+        assert_eq!(events[0].dcc_type, "houdini");
+        assert_eq!(
+            events[0].instance_id,
+            "aaaaaaaa-0000-0000-0000-000000000000"
+        );
+        assert!(
+            events[0]
+                .reason
+                .as_deref()
+                .is_some_and(|reason| reason.contains("request-42"))
+        );
+        let notification = updates
+            .try_recv()
+            .expect("event resource update notification");
+        assert!(notification.contains(GATEWAY_EVENTS_URI));
+    }
+
+    #[tokio::test]
+    async fn gateway_feedback_rejects_empty_required_text() {
+        let gateway = test_gateway_state("1.2.3");
+        let (status, body) = response_json(
+            handle_v1_feedback(
+                State(gateway),
+                Json(json!({
+                    "tool_name": "maya_scene__save",
+                    "intent": "Save the scene",
+                    "blocker": " ",
+                    "severity": "suggestion"
+                })),
+            )
+            .await,
+        )
+        .await;
+
+        assert_eq!(status, StatusCode::BAD_REQUEST);
+        assert_eq!(body["error"]["kind"], "invalid-feedback");
+    }
+}
+
 fn invalid_feedback(message: String) -> Response {
     (
         StatusCode::BAD_REQUEST,
