+++ crates/dcc-mcp-gateway/src/gateway/rest_openapi_feedback.rs
@@ -1,0 +1,85 @@
+//! OpenAPI fragments for gateway-level agent feedback.
+
+use serde_json::{Value, json};
+
+pub(super) fn path_operation() -> Value {
+    let mut operation = super::rest_openapi::post_operation(
+        &["feedback"],
+        "File gateway-level agent feedback",
+        "Records structured feedback even when the referenced DCC instance has already exited. The receipt points to the gateway event resource containing the bounded in-band record.",
+        super::rest_openapi::request_body_ref("GatewayFeedbackReport"),
+        super::rest_openapi::json_response_ref("GatewayFeedbackReceipt"),
+    );
+    let responses = operation["post"]["responses"]
+        .as_object_mut()
+        .expect("POST operation responses must be an object");
+    let response = responses
+        .remove("200")
+        .expect("POST operation must define a success response");
+    responses.insert("201".to_string(), response);
+    operation
+}
+
+pub(super) fn schemas() -> Vec<(&'static str, Value)> {
+    vec![
+        (
+            "GatewayFeedbackReport",
+            json!({
+                "type": "object",
+                "required": ["tool_name", "intent", "blocker", "severity"],
+                "properties": {
+                    "tool_name": {"type": "string", "minLength": 1, "maxLength": 256},
+                    "intent": {"type": "string", "minLength": 1, "maxLength": 4096},
+                    "attempt": {"type": "string", "maxLength": 4096},
+                    "blocker": {"type": "string", "minLength": 1, "maxLength": 4096},
+                    "severity": {"type": "string", "enum": ["blocked", "workaround_found", "suggestion"]},
+                    "dcc_type": {"type": "string", "maxLength": 256},
+                    "instance_id": {"type": "string", "maxLength": 256, "description": "May reference an instance that is no longer live."},
+                    "request_id": {"type": "string", "maxLength": 256},
+                    "job_id": {"type": "string", "maxLength": 256}
+                },
+                "additionalProperties": false,
+            }),
+        ),
+        (
+            "GatewayFeedbackReceipt",
+            json!({
+                "type": "object",
+                "required": ["ok", "success", "feedback_id", "recorded_at", "event_resource_uri"],
+                "properties": {
+                    "ok": {"type": "boolean", "const": true},
+                    "success": {"type": "boolean", "const": true},
+                    "feedback_id": {"type": "string", "format": "uuid"},
+                    "recorded_at": {"type": "string", "format": "date-time"},
+                    "event_resource_uri": {"type": "string", "const": "resources://gateway/events"}
+                },
+                "additionalProperties": false,
+            }),
+        ),
+    ]
+}
+
+#[cfg(test)]
+mod tests {
+    use super::super::rest_openapi::build_gateway_openapi_document;
+
+    #[test]
+    fn documents_feedback_without_live_instance_dependency() {
+        let doc = build_gateway_openapi_document("1.2.3");
+        let operation = &doc["paths"]["/v1/feedback"]["post"];
+        assert_eq!(operation["tags"][0], "feedback");
+        assert!(operation["responses"].get("200").is_none());
+        assert_eq!(
+            operation["responses"]["201"]["content"]["application/json"]["schema"]["$ref"],
+            "#/components/schemas/GatewayFeedbackReceipt"
+        );
+        assert_eq!(
+            operation["requestBody"]["content"]["application/json"]["schema"]["$ref"],
+            "#/components/schemas/GatewayFeedbackReport"
+        );
+        let report = &doc["components"]["schemas"]["GatewayFeedbackReport"];
+        assert!(report["properties"].get("instance_id").is_some());
+        assert!(report["properties"].get("request_id").is_some());
+        assert!(report["properties"].get("job_id").is_some());
+    }
+}
